# optimize(otel): Improve performance of otellog processing with `sync.Pool`