--- a/app/vlinsert/opentelemetry/opentelemetry_test.go
+++ b/app/vlinsert/opentelemetry/opentelemetry_test.go
@@ -207,3 +207,42 @@ func TestPushProtoOk(t *testing.T) {
 func ptrTo[T any](s T) *T {
 	return &s
 }
+
+func BenchmarkPushProto(b *testing.B) {
+	f := func(src []pb.ResourceLogs, timestampsExpected []int64, resultExpected string) {
+		lr := pb.ExportLogsServiceRequest{
+			ResourceLogs: src,
+		}
+
+		pData := lr.MarshalProtobuf(nil)
+		tlp := &insertutil.TestLogMessageProcessor{}
+		if err := pushProtobufRequest(pData, tlp, nil, false); err != nil {
+			b.Fatalf("unexpected error: %s", err)
+		}
+
+		if err := tlp.Verify(timestampsExpected, resultExpected); err != nil {
+			b.Fatal(err)
+		}
+	}
+
+	b.ReportAllocs()
+	b.ResetTimer()
+
+	for b.Loop() {
+		// single line without resource attributes
+		f([]pb.ResourceLogs{
+			{
+				ScopeLogs: []pb.ScopeLogs{
+					{
+						LogRecords: []pb.LogRecord{
+							{Attributes: []*pb.KeyValue{}, TimeUnixNano: 1234, SeverityNumber: 1, Body: pb.AnyValue{StringValue: ptrTo("log-line-message")}},
+						},
+					},
+				},
+			},
+		},
+			[]int64{1234},
+			`{"_msg":"log-line-message","severity":"Trace"}`,
+		)
+	}
+}