### Describe Your Changes

Added two `sync.Pool` to reduce memory allocations

1. `exportLogsReqPool` for  `pb.ExportLogsServiceRequest{}` objects
2. ` fieldsPool` for ` []logstorage.Field slices`
```txt
goos: darwin
goarch: arm64
pkg: github.com/VictoriaMetrics/VictoriaLogs/app/vlinsert/opentelemetry
cpu: Apple M3
            │ before.txt  │             after.txt             │
            │   sec/op    │   sec/op     vs base              │
PushProto-8   711.5n ± 6%   656.0n ± 0%  -7.81% (p=0.002 n=6)

            │ before.txt │             after.txt             │
            │    B/op    │    B/op     vs base               │
PushProto-8   753.0 ± 0%   609.0 ± 0%  -19.12% (p=0.002 n=6)

            │ before.txt │             after.txt             │
            │ allocs/op  │ allocs/op   vs base               │
PushProto-8   21.00 ± 0%   18.00 ± 0%  -14.29% (p=0.002 n=6)
```