@jiekun 2025-09-15
Hello. Thanks for the pull request!

Just curious, what happen if we just use `[]logstorage.Field` instead of `*[]logstorage.Field`:
```go
func getFields() []logstorage.Field {
	if v := fieldsPool.Get(); v != nil {
		return v.([]logstorage.Field)
	}
	return []logstorage.Field{}
}

func putFields(fields []logstorage.Field) {
	fields = fields[:0]
	fieldsPool.Put(fields)
}
```

and change only `var commonFields []logstorage.Field` into `commentFields := getFields()` and `defer putFields(commonFields)`:

<details>

```go
	commonFields := getFields()
	defer putFields(commonFields)

	for _, rl := range req.ResourceLogs {
		commonFields = commonFields[:0]
		commonFields = appendKeyValues(commonFields, rl.Resource.Attributes, "")
		commonFieldsLen := len(commonFields)
		for _, sc := range rl.ScopeLogs {
			commonFields = pushFieldsFromScopeLogs(&sc, commonFields[:commonFieldsLen], lmp, msgFields, useDefaultStreamFields)
		}
	}
```

</details>

By using a pointer to a slice, the underlying array of the slice will still be replaced when its cap (capacity) is about to be exceeded. So I'm curious about the benefits of using pointers here, and it would be even better if comparison result could be provided!