@yumosx 2025-09-15
hi, @jiekun return a pointer to avoid extra allocation on `Put()`