@func25 2025-09-16
> hi, @jiekun return a pointer to avoid extra allocation on Put()

Mentioned in an old [Go article](https://victoriametrics.com/blog/go-sync-pool/#syncpool-and-allocation-trap) ;)